Solve the template lookup problem in tstream.
* Bug: In Hurricane, in tstream (Commons.h), in the *template* overload of
    operator<<(), do not use the operator<<() of ostream as it will be
    looked up in "stage 1" (template definition) and so will miss all the
    overloads added later and built over getString<>(). Instead, make use
    of getString<>(), which, as another template will be looked up in
    "stage 2" (template instanciation) and at that point will have all the
    needed template specialisation of getString<>().
      We also need to define new stream manipulators to be able to create
    a matching template overload not dependant from the implementation.
    To avoid name clashes, we prefix a 't'. For now, only 'tsetw()' is
    refined.
      As a side effect, we cannot directly print bit-fields into the stream,
    we must go through an intermediate variable (happens once in AutoContact).

+++ katabatic/src/AutoContact.cpp
@@ -325,7 +325,11 @@ namespace Katabatic {
       _dyMin = 0;
       _dxMax = (int)DbU::toLambda( _gcell->getXMax()-_gcell->getX() );
       _dyMax = (int)DbU::toLambda( _gcell->getYMax()-_gcell->getY() );
-      cdebug_log(145,0) << "* deltas: [" << _dxMin << " " << _dyMin << " " << _dxMax << " " << _dyMax << "]" << endl;
+      cdebug_log(145,0) << "* deltas: [" << (int)_dxMin
+                        << " "           << (int)_dyMin
+                        << " "           << (int)_dxMax
+                        << " "           << (int)_dyMax
+                        << "]" << endl;
     } else {
       cerr << Bug( "NULL GCell for %s.", _getString().c_str() ) << endl;
     }
